Add formatExposureRange method, rename formatStatus file
